reorganised project into Main, Naruto and Printutil
