+++ src/main/java/Naruto.java
@@ -1,42 +1,15 @@
-import java.util.Scanner;
-
 public class Naruto {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        greet();
-        while (true) {
-            String s = sc.nextLine();
-            if (s.equals("bye")) {
-                exit();
-            }
-            echo(s);
-        }
-    }
-
+    Naruto() {}
     static void greet() {
-        spacer();
-        System.out.println("    Hello! I'm Naruto, and I'm gonna become the next Hokage!" +
+        PrintUtil.print("Hello! I'm Naruto, and I'm gonna become a ninja!" +
                 "\n    Believe it!");
-        spacer();
-
     }
-
-    static void exit() {
-        spacer();
-        System.out.println("    See you next time!");
-        spacer();
-        System.exit(0);
-    }
-
-    static void spacer() {
-        System.out.println("    --------------------------------------------------------" +
-                "----------");
+    public static void sayBye() {
+        PrintUtil.print("Bye bye!");
     }
 
     static void echo(String s) {
-        spacer();
-        System.out.println("    " + s);
-        spacer();
+        PrintUtil.print(s);
     }
 
 }
